Add tracing timeline span description
