+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/index.js
@@ -20,9 +20,20 @@ import "./index.css";
 import App from "./App";
 import React from "react";
 import ReactDOM from "react-dom";
+import {MuiThemeProvider, createMuiTheme} from "@material-ui/core/styles";
 import * as serviceWorker from "./serviceWorker";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const theme = createMuiTheme({
+    typography: {
+        useNextVariants: true
+    }
+});
+
+ReactDOM.render((
+    <MuiThemeProvider theme={theme}>
+        <App/>
+    </MuiThemeProvider>
+), document.getElementById("root"));
 
 /*
  * If you want your app to work offline and load faster, you can change
