+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/pages/tracing/view/SequenceDiagram.js
@@ -16,22 +16,10 @@
  * under the License.
  */
 
-import PropTypes from "prop-types";
 import React from "react";
-import Search from "./Search";
-import View from "./view";
-import {Redirect, Route, Switch, withRouter} from "react-router-dom";
 
-const Tracing = ({match}) => (
-    <Switch>
-        <Route exact path={`${match.url}/search`} component={Search}/>
-        <Route exact path={`${match.url}/id/:traceId`} component={View}/>
-        <Redirect from={`${match.url}/`} to={`${match.url}/search`}/>
-    </Switch>
+const SequenceDiagram = () => (
+    <div>Sequence Diagram</div>
 );
 
-Tracing.propTypes = {
-    match: PropTypes.object.isRequired
-};
-
-export default withRouter(Tracing);
+export default SequenceDiagram;
